Restructured project a bit so its clearer to read and undestand
